Moved matcher related events and actions to extension package.
