+++ webofneeds/won-bot/src/main/java/won/bot/framework/extensions/matcher/MatcherExtensionAtomCreatedEvent.java
@@ -8,17 +8,26 @@
  * KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-package won.bot.framework.eventbot.event.impl.matcher;
+package won.bot.framework.extensions.matcher;
 
 import java.net.URI;
+
+import org.apache.jena.query.Dataset;
 
 import won.bot.framework.eventbot.event.BaseAtomSpecificEvent;
 
 /**
  *
  */
-public class MatcherExtensionAtomDeactivatedEvent extends BaseAtomSpecificEvent {
-    public MatcherExtensionAtomDeactivatedEvent(final URI atomURI) {
+public class MatcherExtensionAtomCreatedEvent extends BaseAtomSpecificEvent {
+    private final Dataset atomData;
+
+    public MatcherExtensionAtomCreatedEvent(final URI atomURI, final Dataset atomData) {
         super(atomURI);
+        this.atomData = atomData;
+    }
+
+    public Dataset getAtomData() {
+        return atomData;
     }
 }
