+++ webofneeds/won-bot/src/main/java/won/bot/framework/extensions/matcher/MatcherExtensionRegisterSucceededEvent.java
@@ -8,26 +8,17 @@
  * KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-package won.bot.framework.eventbot.event.impl.matcher;
+package won.bot.framework.extensions.matcher;
 
 import java.net.URI;
 
-import org.apache.jena.query.Dataset;
-
-import won.bot.framework.eventbot.event.BaseAtomSpecificEvent;
+import won.bot.framework.eventbot.event.BaseNodeSpecificEvent;
 
 /**
  *
  */
-public class MatcherExtensionAtomCreatedEvent extends BaseAtomSpecificEvent {
-    private final Dataset atomData;
-
-    public MatcherExtensionAtomCreatedEvent(final URI atomURI, final Dataset atomData) {
-        super(atomURI);
-        this.atomData = atomData;
-    }
-
-    public Dataset getAtomData() {
-        return atomData;
+public class MatcherExtensionRegisterSucceededEvent extends BaseNodeSpecificEvent {
+    public MatcherExtensionRegisterSucceededEvent(final URI wonNodeURI) {
+        super(wonNodeURI);
     }
 }
